[sw/silicon_creator] Fix type issues on abs_mmio
Fix type issues on `abs_mmio_read8()` and `abs_mmio_write8()` functions.

+++ sw/device/silicon_creator/lib/base/abs_mmio.c
@@ -6,7 +6,7 @@
 
 // `extern` declarations to give the inline functions in the corresponding
 // header a link location.
-extern uint32_t abs_mmio_read8(uint32_t addr);
+extern uint8_t abs_mmio_read8(uint32_t addr);
 extern void abs_mmio_write8(uint32_t addr, uint8_t value);
 extern uint32_t abs_mmio_read32(uint32_t addr);
 extern void abs_mmio_write32(uint32_t addr, uint32_t value);
